Fix: decode uv_index as Double instead of Int to match Open-Meteo API
Open-Meteo returns uv_index as a floating-point value (e.g. 3.75).
Decoding it as Int caused a typeMismatch DecodingError, surfaced to the
user as "Veriler doğru biçimde olmadığı için okunamadı."

Changed Current.uvIndex, Hourly.uvIndex, and Daily.uvIndexMax from Int
to Double in OpenMeteoModels. WeatherMapping already converts to Int at
mapping time; current-weather path updated to use Int(rounded()).

+++ Praeventus.swiftpm/OpenMeteoModels.swift
@@ -28,7 +28,7 @@ struct ForecastResponse: Decodable, Equatable {
         let windGusts10m: Double?
         let precipitationProbability: Double?
         let weatherCode: Int?
-        let uvIndex: Int?
+        let uvIndex: Double?
         let dewPoint2m: Double?
         let visibility: Double?
 
@@ -55,7 +55,7 @@ struct ForecastResponse: Decodable, Equatable {
         let temperature2m: [Double?]?
         let precipitationProbability: [Double?]?
         let weatherCode: [Int?]?
-        let uvIndex: [Int?]?
+        let uvIndex: [Double?]?
         let windSpeed10m: [Double?]?
         let windDirection10m: [Int?]?
         let windGusts10m: [Double?]?
@@ -84,7 +84,7 @@ struct ForecastResponse: Decodable, Equatable {
         let temperature2mMin: [Double?]?
         let apparentTemperatureMax: [Double?]?
         let apparentTemperatureMin: [Double?]?
-        let uvIndexMax: [Int?]?
+        let uvIndexMax: [Double?]?
         let windSpeed10mMax: [Double?]?
         let windDirection10mDominant: [Int?]?
         let windGusts10mMax: [Double?]?
